fix: Correct index order when referencing 2D list
In order to reference the correct position in the 2D list (when
displayed as a board/grid) the x and y coordinates need to be swapped.
This is due to the first index position specified (list[i]) representing
the y coordinate (or the row) and the second index position (list[i][j])
representing the x coordinate (or the column). This was missed in the
initial testing but is corrected by this commit.