fix: Remove validation check in fire_missile function
The fire_missile function contained an unnecessary call to the
check_valid_position function leftover from earlier development.

When the code to add guesses to the opponents play board was introduced
(commit 009a4b8), this validation check produced errors as it did not
contain logic to handle the new strings "Hit" and "Miss" added to the
the play_board list, therefore the check returned False.

A string would then be checked to see if it was 'in' the returned value
raising a TypeError as a bool is not iterable.